[REACTOS] Improve <imm.h> and <immdev.h> compatibility
