+++ dll/ime/msctfime/msctfime.h
@@ -15,7 +15,8 @@
 
 #include <windows.h>
 #include <imm.h>
-#include <ddk/immdev.h>
+#include <immdev.h>
+#include <imm32_undoc.h>
 #include <cguid.h>
 #include <tchar.h>
 #include <msctf.h>
@@ -33,7 +34,5 @@
 #include <wine/debug.h>
 
 #include "resource.h"
-
-#define IS_IME_HKL(hKL) ((((ULONG_PTR)(hKL)) & 0xF0000000) == 0xE0000000)
 
 extern HINSTANCE g_hInst;
