+++ dll/win32/imm32/precomp.h
@@ -27,7 +27,8 @@
 #include <winver.h>
 
 #include <imm.h>
-#include <ddk/immdev.h>
+#include <immdev.h>
+#include <imm32_undoc.h>
 
 #define NTOS_MODE_USER
 #include <ndk/umtypes.h>
@@ -40,7 +41,6 @@
 
 /* Undocumented user definitions */
 #include <undocuser.h>
-#include <imm32_undoc.h>
 
 #include <strsafe.h>
