+++ dll/win32/msctf/utils.cpp
@@ -16,7 +16,7 @@
 #include <windows.h>
 #include <sddl.h>
 #include <imm.h>
-#include <ddk/immdev.h>
+#include <immdev.h>
 #include <cguid.h>
 #include <tchar.h>
 #include <msctf.h>
