+++ dll/win32/msutb/precomp.h
@@ -16,7 +16,7 @@
 #include <windows.h>
 #include <oleacc.h>
 #include <imm.h>
-#include <ddk/immdev.h>
+#include <immdev.h>
 #include <undocuser.h>
 #include <cguid.h>
 #include <msctf.h>
