+++ sdk/include/psdk/imm.h
@@ -25,34 +25,34 @@
 extern "C" {
 #endif
 
-#ifdef __REACTOS__
+#if (WINVER >= 0x040A)
 DECLARE_HANDLE(HIMC);
 DECLARE_HANDLE(HIMCC);
-typedef HKL *LPHKL;
 #else
-typedef HANDLE HIMC;
-typedef HANDLE HIMCC;
+typedef DWORD HIMC;
+typedef DWORD HIMCC;
 #endif
+
+typedef HKL FAR *LPHKL;
+typedef UINT FAR *LPUINT;
 
 typedef struct tagREGISTERWORDA {
     LPSTR lpReading;
     LPSTR lpWord;
-} REGISTERWORDA, *PREGISTERWORDA, *NPREGISTERWORDA, *LPREGISTERWORDA;
+} REGISTERWORDA, *PREGISTERWORDA, NEAR *NPREGISTERWORDA, FAR *LPREGISTERWORDA;
 
 typedef struct tagREGISTERWORDW {
     LPWSTR lpReading;
     LPWSTR lpWord;
-} REGISTERWORDW, *PREGISTERWORDW, *NPREGISTERWORDW, *LPREGISTERWORDW;
+} REGISTERWORDW, *PREGISTERWORDW, NEAR *NPREGISTERWORDW, FAR *LPREGISTERWORDW;
 
-typedef int
-(CALLBACK *REGISTERWORDENUMPROCA)(
+typedef INT (CALLBACK *REGISTERWORDENUMPROCA)(
   _In_ LPCSTR,
   _In_ DWORD,
   _In_ LPCSTR,
   _In_ LPVOID);
 
-typedef int
-(CALLBACK *REGISTERWORDENUMPROCW)(
+typedef INT (CALLBACK *REGISTERWORDENUMPROCW)(
   _In_ LPCWSTR,
   _In_ DWORD,
   _In_ LPCWSTR,
@@ -60,81 +60,90 @@ typedef int
 
 typedef struct tagCANDIDATEFORM
 {
-				DWORD   dwIndex;
-				DWORD   dwStyle;
-				POINT ptCurrentPos;
-				RECT  rcArea;
-} CANDIDATEFORM, *LPCANDIDATEFORM;
+    DWORD dwIndex;
+    DWORD dwStyle;
+    POINT ptCurrentPos;
+    RECT  rcArea;
+} CANDIDATEFORM, *PCANDIDATEFORM, NEAR *NPCANDIDATEFORM, FAR *LPCANDIDATEFORM;
 
 typedef struct tagCANDIDATELIST
 {
-				DWORD dwSize;
-				DWORD dwStyle;
-				DWORD dwCount;
-				DWORD dwSelection;
-				DWORD dwPageStart;
-				DWORD dwPageSize;
-				DWORD dwOffset[1];
-} CANDIDATELIST, *LPCANDIDATELIST;
+    DWORD dwSize;
+    DWORD dwStyle;
+    DWORD dwCount;
+    DWORD dwSelection;
+    DWORD dwPageStart;
+    DWORD dwPageSize;
+    DWORD dwOffset[1];
+} CANDIDATELIST, *PCANDIDATELIST, NEAR *NPCANDIDATELIST, FAR *LPCANDIDATELIST;
+
+#define STYLE_DESCRIPTION_SIZE 32
 
 typedef struct tagSTYLEBUFA
 {
-				DWORD dwStyle;
-				CHAR  szDescription[32];
-} STYLEBUFA, *LPSTYLEBUFA;
+    DWORD dwStyle;
+    CHAR  szDescription[STYLE_DESCRIPTION_SIZE];
+} STYLEBUFA, *PSTYLEBUFA, NEAR *NPSTYLEBUFA, FAR *LPSTYLEBUFA;
 
 typedef struct tagSTYLEBUFW
 {
-				DWORD dwStyle;
-				WCHAR  szDescription[32];
-} STYLEBUFW, *LPSTYLEBUFW;
+    DWORD dwStyle;
+    WCHAR  szDescription[STYLE_DESCRIPTION_SIZE];
+} STYLEBUFW, *PSTYLEBUFW, NEAR *NPSTYLEBUFW, FAR *LPSTYLEBUFW;
 
-DECL_WINELIB_TYPE_AW(STYLEBUF)
-DECL_WINELIB_TYPE_AW(LPSTYLEBUF)
+#if (WINVER >= 0x040A)
 
 typedef struct tagRECONVERTSTRING
 {
-	DWORD	dwSize;
-	DWORD	dwVersion;
-	DWORD	dwStrLen;
-	DWORD	dwStrOffset;
-	DWORD	dwCompStrLen;
-	DWORD	dwCompStrOffset;
-	DWORD	dwTargetStrLen;
-	DWORD	dwTargetStrOffset;
-} RECONVERTSTRING, *LPRECONVERTSTRING;
-
+    DWORD dwSize;
+    DWORD dwVersion;
+    DWORD dwStrLen;
+    DWORD dwStrOffset;
+    DWORD dwCompStrLen;
+    DWORD dwCompStrOffset;
+    DWORD dwTargetStrLen;
+    DWORD dwTargetStrOffset;
+} RECONVERTSTRING, *PRECONVERTSTRING, NEAR *NPRECONVERTSTRING, FAR *LPRECONVERTSTRING;
 
 #define IMEMENUITEM_STRING_SIZE 80
 
 typedef struct tagIMEMENUITEMINFOA
 {
-	UINT		cbSize;
-	UINT		fType;
-	UINT		fState;
-	UINT		wID;
-	HBITMAP		hbmpChecked;
-	HBITMAP		hbmpUnchecked;
-	DWORD		dwItemData;
-	CHAR		szString[IMEMENUITEM_STRING_SIZE];
-	HBITMAP		hbmpItem;
-} IMEMENUITEMINFOA, *LPIMEMENUITEMINFOA;
+    UINT        cbSize;
+    UINT        fType;
+    UINT        fState;
+    UINT        wID;
+    HBITMAP     hbmpChecked;
+    HBITMAP     hbmpUnchecked;
+    DWORD       dwItemData;
+    CHAR        szString[IMEMENUITEM_STRING_SIZE];
+    HBITMAP     hbmpItem;
+} IMEMENUITEMINFOA, *PIMEMENUITEMINFOA, NEAR *NPIMEMENUITEMINFOA, FAR *LPIMEMENUITEMINFOA;
 
 typedef struct tagIMEMENUITEMINFOW
 {
-	UINT		cbSize;
-	UINT		fType;
-	UINT		fState;
-	UINT		wID;
-	HBITMAP		hbmpChecked;
-	HBITMAP		hbmpUnchecked;
-	DWORD		dwItemData;
-	WCHAR		szString[IMEMENUITEM_STRING_SIZE];
-	HBITMAP		hbmpItem;
-} IMEMENUITEMINFOW, *LPIMEMENUITEMINFOW;
+    UINT        cbSize;
+    UINT        fType;
+    UINT        fState;
+    UINT        wID;
+    HBITMAP     hbmpChecked;
+    HBITMAP     hbmpUnchecked;
+    DWORD       dwItemData;
+    WCHAR       szString[IMEMENUITEM_STRING_SIZE];
+    HBITMAP     hbmpItem;
+} IMEMENUITEMINFOW, *PIMEMENUITEMINFOW, NEAR *NPIMEMENUITEMINFOW, FAR *LPIMEMENUITEMINFOW;
 
-DECL_WINELIB_TYPE_AW(IMEMENUITEMINFO)
-DECL_WINELIB_TYPE_AW(LPIMEMENUITEMINFO)
+#ifdef UNICODE
+    typedef IMEMENUITEMINFOW IMEMENUITEMINFO;
+    typedef PIMEMENUITEMINFOW PIMEMENUITEMINFO;
+    typedef NPIMEMENUITEMINFOW NPIMEMENUITEMINFO;
+    typedef LPIMEMENUITEMINFOW LPIMEMENUITEMINFO;
+#else
+    typedef IMEMENUITEMINFOA IMEMENUITEMINFO;
+    typedef PIMEMENUITEMINFOA PIMEMENUITEMINFO;
+    typedef NPIMEMENUITEMINFOA NPIMEMENUITEMINFO;
+    typedef LPIMEMENUITEMINFOA LPIMEMENUITEMINFO;
+#endif
 
 typedef struct tagIMECHARPOSITION {
     DWORD       dwSize;
@@ -144,131 +153,16 @@ typedef struct tagIMECHARPOSITION {
     RECT        rcDocument;
 } IMECHARPOSITION, *PIMECHARPOSITION, *NPIMECHARPOSITION, *LPIMECHARPOSITION;
 
-typedef BOOL    (CALLBACK* IMCENUMPROC)(HIMC, LPARAM);
+typedef BOOL (CALLBACK* IMCENUMPROC)(HIMC, LPARAM);
+
+#endif // (WINVER >= 0x040A)
 
 typedef struct _tagCOMPOSITIONFORM
 {
-				DWORD   dwStyle;
-				POINT ptCurrentPos;
-				RECT  rcArea;
-} COMPOSITIONFORM, *LPCOMPOSITIONFORM;
-
-typedef struct _tagIMEINFO {
-    DWORD       dwPrivateDataSize;
-    DWORD       fdwProperty;
-    DWORD       fdwConversionCaps;
-    DWORD       fdwSentenceCaps;
-    DWORD       fdwUICaps;
-    DWORD       fdwSCSCaps;
-    DWORD       fdwSelectCaps;
-} IMEINFO, *LPIMEINFO;
-
-typedef struct tagCOMPOSITIONSTRING {
-    DWORD dwSize;
-    DWORD dwCompReadAttrLen;
-    DWORD dwCompReadAttrOffset;
-    DWORD dwCompReadClauseLen;
-    DWORD dwCompReadClauseOffset;
-    DWORD dwCompReadStrLen;
-    DWORD dwCompReadStrOffset;
-    DWORD dwCompAttrLen;
-    DWORD dwCompAttrOffset;
-    DWORD dwCompClauseLen;
-    DWORD dwCompClauseOffset;
-    DWORD dwCompStrLen;
-    DWORD dwCompStrOffset;
-    DWORD dwCursorPos;
-    DWORD dwDeltaStart;
-    DWORD dwResultReadClauseLen;
-    DWORD dwResultReadClauseOffset;
-    DWORD dwResultReadStrLen;
-    DWORD dwResultReadStrOffset;
-    DWORD dwResultClauseLen;
-    DWORD dwResultClauseOffset;
-    DWORD dwResultStrLen;
-    DWORD dwResultStrOffset;
-    DWORD dwPrivateSize;
-    DWORD dwPrivateOffset;
-} COMPOSITIONSTRING, *LPCOMPOSITIONSTRING;
-
-typedef struct tagGUIDELINE {
-    DWORD dwSize;
-    DWORD dwLevel;
-    DWORD dwIndex;
-    DWORD dwStrLen;
-    DWORD dwStrOffset;
-    DWORD dwPrivateSize;
-    DWORD dwPrivateOffset;
-} GUIDELINE, *LPGUIDELINE;
-
-typedef struct tagCANDIDATEINFO {
-    DWORD               dwSize;
-    DWORD               dwCount;
-    DWORD               dwOffset[32];
-    DWORD               dwPrivateSize;
-    DWORD               dwPrivateOffset;
-} CANDIDATEINFO, *LPCANDIDATEINFO;
-
-/* IME Property bits */
-#define IME_PROP_END_UNLOAD             0x0001
-#define IME_PROP_KBD_CHAR_FIRST         0x0002
-#define IME_PROP_IGNORE_UPKEYS          0x0004
-#define IME_PROP_NEED_ALTKEY            0x0008
-#define IME_PROP_NO_KEYS_ON_CLOSE       0x0010
-#define IME_PROP_ACCEPT_WIDE_VKEY       0x0020
-
-/* for NI_CONTEXTUPDATED */
-#define IMC_SETCONVERSIONMODE           0x0002
-#define IMC_SETSENTENCEMODE             0x0004
-#define IMC_SETOPENSTATUS               0x0006
-
-/* dwAction for ImmNotifyIME */
-#define NI_CONTEXTUPDATED               0x0003
-#define NI_OPENCANDIDATE                0x0010
-#define NI_CLOSECANDIDATE               0x0011
-#define NI_SELECTCANDIDATESTR           0x0012
-#define NI_CHANGECANDIDATELIST          0x0013
-#define NI_FINALIZECONVERSIONRESULT     0x0014
-#define NI_COMPOSITIONSTR               0x0015
-#define NI_SETCANDIDATE_PAGESTART       0x0016
-#define NI_SETCANDIDATE_PAGESIZE        0x0017
-#define NI_IMEMENUSELECTED              0x0018
-
-/* dwSystemInfoFlags bits */
-#define IME_SYSINFO_WINLOGON            0x0001
-#define IME_SYSINFO_WOW16               0x0002
-
-BOOL  WINAPI ImmUnlockIMC(HIMC);
-DWORD WINAPI ImmGetIMCLockCount(HIMC);
-HIMCC  WINAPI ImmCreateIMCC(DWORD);
-HIMCC  WINAPI ImmDestroyIMCC(HIMCC);
-LPVOID WINAPI ImmLockIMCC(HIMCC);
-BOOL   WINAPI ImmUnlockIMCC(HIMCC);
-DWORD  WINAPI ImmGetIMCCLockCount(HIMCC);
-HIMCC  WINAPI ImmReSizeIMCC(HIMCC, DWORD);
-DWORD  WINAPI ImmGetIMCCSize(HIMCC);
-
-BOOL WINAPI ImmGenerateMessage(HIMC);
-LRESULT WINAPI ImmRequestMessageA(HIMC, WPARAM, LPARAM);
-LRESULT WINAPI ImmRequestMessageW(HIMC, WPARAM, LPARAM);
-#define ImmRequestMessage WINELIB_NAME_AW(ImmRequestMessage);
-BOOL WINAPI ImmTranslateMessage(HWND, UINT, WPARAM, LPARAM);
-
-HWND WINAPI
-ImmCreateSoftKeyboard(
-    _In_ UINT uType,
-    _In_ HWND hwndParent,
-    _In_ INT x,
-    _In_ INT y);
-
-BOOL WINAPI
-ImmShowSoftKeyboard(
-    _In_ HWND hwndSoftKBD,
-    _In_ INT nCmdShow);
-
-BOOL WINAPI
-ImmDestroySoftKeyboard(
-    _In_ HWND hwndSoftKBD);
+    DWORD   dwStyle;
+    POINT ptCurrentPos;
+    RECT  rcArea;
+} COMPOSITIONFORM, *PCOMPOSITIONFORM, NEAR *NPCOMPOSITIONFORM, FAR *LPCOMPOSITIONFORM;
 
 /* wParam for WM_IME_CONTROL */
 #define IMC_GETCANDIDATEPOS             0x0007
@@ -294,7 +188,7 @@ ImmDestroySoftKeyboard(
 #define NI_COMPOSITIONSTR               0x0015
 #define NI_SETCANDIDATE_PAGESTART       0x0016
 #define NI_SETCANDIDATE_PAGESIZE        0x0017
-#define NI_IMEMENUSELECTED		0x0018
+#define NI_IMEMENUSELECTED              0x0018
 
 /* lParam for WM_IME_SETCONTEXT */
 #define ISC_SHOWUICANDIDATEWINDOW       0x00000001
@@ -303,13 +197,11 @@ ImmDestroySoftKeyboard(
 #define ISC_SHOWUIALLCANDIDATEWINDOW    0x0000000F
 #define ISC_SHOWUIALL                   0xC000000F
 
-
 /* dwIndex for ImmNotifyIME/NI_COMPOSITIONSTR */
 #define CPS_COMPLETE                    0x0001
 #define CPS_CONVERT                     0x0002
 #define CPS_REVERT                      0x0003
 #define CPS_CANCEL                      0x0004
-
 
 /* the modifiers of hot key */
 #define MOD_ALT                         0x0001
@@ -321,7 +213,6 @@ ImmDestroySoftKeyboard(
 
 #define MOD_ON_KEYUP                    0x0800
 #define MOD_IGNORE_ALL_MODIFIER         0x0400
-
 
 /* Windows for Simplified Chinese Edition hot key ID from 0x10 - 0x2F */
 #define IME_CHOTKEY_IME_NONIME_TOGGLE           0x10
@@ -349,14 +240,10 @@ ImmDestroySoftKeyboard(
 #define IME_ITHOTKEY_RESEND_RESULTSTR           0x200
 #define IME_ITHOTKEY_PREVIOUS_COMPOSITION       0x201
 #define IME_ITHOTKEY_UISTYLE_TOGGLE             0x202
-#define IME_ITHOTKEY_RECONVERTSTRING		0x203
+#define IME_ITHOTKEY_RECONVERTSTRING            0x203
 #define IME_HOTKEY_PRIVATE_LAST                 0x21F
 
-
-/*
- * dwSystemInfoFlags bits
- * parameter of ImmGetCompositionString
- */
+/* parameter of ImmGetCompositionString */
 #define GCS_COMPREADSTR                 0x0001
 #define GCS_COMPREADATTR                0x0002
 #define GCS_COMPREADCLAUSE              0x0004
@@ -374,11 +261,7 @@ ImmDestroySoftKeyboard(
 #define CS_INSERTCHAR                   0x2000
 #define CS_NOMOVECARET                  0x4000
 
-
-/*
- * bits of fdwInit of INPUTCONTEXT
- * IME version constants
- */
+/* IME version constants */
 #define IMEVER_0310                     0x0003000A
 #define IMEVER_0400                     0x00040000
 
@@ -387,7 +270,7 @@ ImmDestroySoftKeyboard(
 #define IME_PROP_SPECIAL_UI             0x00020000
 #define IME_PROP_CANDLIST_START_FROM_1  0x00040000
 #define IME_PROP_UNICODE                0x00080000
-#define IME_PROP_COMPLETE_ON_UNSELECT	0x00100000
+#define IME_PROP_COMPLETE_ON_UNSELECT   0x00100000
 
 /* IME UICapability bits */
 #define UI_CAP_2700                     0x00000001
@@ -398,12 +281,11 @@ ImmDestroySoftKeyboard(
 /* ImmSetCompositionString Capability bits */
 #define SCS_CAP_COMPSTR                 0x00000001
 #define SCS_CAP_MAKEREAD                0x00000002
-#define SCS_CAP_SETRECONVERTSTRING	0x00000004
+#define SCS_CAP_SETRECONVERTSTRING      0x00000004
 
 /* IME WM_IME_SELECT inheritance Capability bits */
 #define SELECT_CAP_CONVERSION           0x00000001
 #define SELECT_CAP_SENTENCE             0x00000002
-
 
 /* ID for deIndex of ImmGetGuideLine */
 #define GGL_LEVEL                       0x00000001
@@ -445,13 +327,12 @@ ImmDestroySoftKeyboard(
 #define IGP_SETCOMPSTR                  0x00000014
 #define IGP_SELECT                      0x00000018
 
-
 /* dwIndex for ImmSetCompositionString API */
-#define SCS_SETSTR          (GCS_COMPREADSTR|GCS_COMPSTR)
-#define SCS_CHANGEATTR      (GCS_COMPREADATTR|GCS_COMPATTR)
-#define SCS_CHANGECLAUSE    (GCS_COMPREADCLAUSE|GCS_COMPCLAUSE)
-#define SCS_SETRECONVERTSTRING		0x00010000
-#define SCS_QUERYRECONVERTSTRING	0x00020000
+#define SCS_SETSTR                      (GCS_COMPREADSTR|GCS_COMPSTR)
+#define SCS_CHANGEATTR                  (GCS_COMPREADATTR|GCS_COMPATTR)
+#define SCS_CHANGECLAUSE                (GCS_COMPREADCLAUSE|GCS_COMPCLAUSE)
+#define SCS_SETRECONVERTSTRING          0x00010000
+#define SCS_QUERYRECONVERTSTRING        0x00020000
 
 /* attribute for COMPOSITIONSTRING Structure */
 #define ATTR_INPUT                      0x00
@@ -459,8 +340,7 @@ ImmDestroySoftKeyboard(
 #define ATTR_CONVERTED                  0x02
 #define ATTR_TARGET_NOTCONVERTED        0x03
 #define ATTR_INPUT_ERROR                0x04
-#define ATTR_FIXEDCONVERTED		0x05
-
+#define ATTR_FIXEDCONVERTED             0x05
 
 /* bit field for IMC_SETCOMPOSITIONWINDOW, IMC_SETCANDIDATEWINDOW */
 #define CFS_DEFAULT                     0x0000
@@ -474,7 +354,6 @@ ImmDestroySoftKeyboard(
 #define GCL_CONVERSION                  0x0001
 #define GCL_REVERSECONVERSION           0x0002
 #define GCL_REVERSE_LENGTH              0x0003
-
 
 /* bit field for conversion mode */
 #define IME_CMODE_ALPHANUMERIC          0x0000
@@ -494,15 +373,15 @@ ImmDestroySoftKeyboard(
 #define IME_CMODE_NOCONVERSION          0x0100
 #define IME_CMODE_EUDC                  0x0200
 #define IME_CMODE_SYMBOL                0x0400
-#define IME_CMODE_FIXED			0x0800
+#define IME_CMODE_FIXED                 0x0800
 
-
+/* bit field for sentence mode */
 #define IME_SMODE_NONE                  0x0000
 #define IME_SMODE_PLAURALCLAUSE         0x0001
 #define IME_SMODE_SINGLECONVERT         0x0002
 #define IME_SMODE_AUTOMATIC             0x0004
 #define IME_SMODE_PHRASEPREDICT         0x0008
-#define IME_SMODE_CONVERSATION		0x0010
+#define IME_SMODE_CONVERSATION          0x0010
 
 /* style of candidate */
 #define IME_CAND_UNKNOWN                0x0000
@@ -511,7 +390,6 @@ ImmDestroySoftKeyboard(
 #define IME_CAND_MEANING                0x0003
 #define IME_CAND_RADICAL                0x0004
 #define IME_CAND_STROKE                 0x0005
-
 
 /* wParam of report message WM_IME_NOTIFY */
 #define IMN_CLOSESTATUSWINDOW           0x0001
@@ -529,21 +407,22 @@ ImmDestroySoftKeyboard(
 #define IMN_GUIDELINE                   0x000D
 #define IMN_PRIVATE                     0x000E
 
+#if (WINVER >= 0x040A)
 
 /* wParam of report message WM_IME_REQUEST */
-#define IMR_COMPOSITIONWINDOW		0x0001
-#define IMR_CANDIDATEWINDOW		0x0002
-#define IMR_COMPOSITIONFONT		0x0003
-#define IMR_RECONVERTSTRING		0x0004
-#define IMR_CONFIRMRECONVERTSTRING	0x0005
-#define IMR_QUERYCHARPOSITION		0x0006
-#define IMR_DOCUMENTFEED		0x0007
+#define IMR_COMPOSITIONWINDOW           0x0001
+#define IMR_CANDIDATEWINDOW             0x0002
+#define IMR_COMPOSITIONFONT             0x0003
+#define IMR_RECONVERTSTRING             0x0004
+#define IMR_CONFIRMRECONVERTSTRING      0x0005
+#define IMR_QUERYCHARPOSITION           0x0006
+#define IMR_DOCUMENTFEED                0x0007
 
+#endif
 
 /* error code of ImmGetCompositionString */
 #define IMM_ERROR_NODATA                (-1)
 #define IMM_ERROR_GENERAL               (-2)
-
 
 /* dialog mode of ImmConfigureIME */
 #define IME_CONFIG_GENERAL              1
@@ -565,13 +444,14 @@ ImmDestroySoftKeyboard(
 #define IME_ESC_HANJA_MODE              0x1008
 #define IME_ESC_AUTOMATA                0x1009
 #define IME_ESC_PRIVATE_HOTKEY          0x100a
-#define IME_ESC_GETHELPFILENAME		0x100b
-
+#define IME_ESC_GETHELPFILENAME         0x100b
 
 /* style of word registration */
 #define IME_REGWORD_STYLE_EUDC          0x00000001
 #define IME_REGWORD_STYLE_USER_FIRST    0x80000000
 #define IME_REGWORD_STYLE_USER_LAST     0xFFFFFFFF
+
+#if (WINVER >= 0x040A)
 
 /* dwFlags for ImmAssociateContextEx */
 #define IACE_CHILDREN           0x0001
@@ -579,32 +459,33 @@ ImmDestroySoftKeyboard(
 #define IACE_IGNORENOCONTEXT    0x0020
 
 /* dwFlags for ImmGetImeMenuItems */
-#define IGIMIF_RIGHTMENU		0x0001
+#define IGIMIF_RIGHTMENU        0x0001
 
 /* dwType for ImmGetImeMenuItems */
-#define IGIMII_CMODE			0x0001
-#define IGIMII_SMODE			0x0002
-#define IGIMII_CONFIGURE		0x0004
-#define IGIMII_TOOLS			0x0008
-#define IGIMII_HELP			0x0010
-#define IGIMII_OTHER			0x0020
-#define IGIMII_INPUTTOOLS		0x0040
+#define IGIMII_CMODE            0x0001
+#define IGIMII_SMODE            0x0002
+#define IGIMII_CONFIGURE        0x0004
+#define IGIMII_TOOLS            0x0008
+#define IGIMII_HELP             0x0010
+#define IGIMII_OTHER            0x0020
+#define IGIMII_INPUTTOOLS       0x0040
 
 /* fType of IMEMENUITEMINFO structure */
-#define IMFT_RADIOCHECK			0x00001
-#define IMFT_SEPARATOR			0x00002
-#define IMFT_SUBMENU			0x00004
+#define IMFT_RADIOCHECK         0x00001
+#define IMFT_SEPARATOR          0x00002
+#define IMFT_SUBMENU            0x00004
 
 /* fState of IMEMENUITEMINFO structure */
-#define IMFS_GRAYED			MFS_GRAYED
-#define IMFS_DISABLED			MFS_DISABLED
-#define IMFS_CHECKED			MFS_CHECKED
-#define IMFS_HILITE			MFS_HILITE
-#define IMFS_ENABLED			MFS_ENABLED
-#define IMFS_UNCHECKED			MFS_UNCHECKED
-#define IMFS_UNHILITE			MFS_UNHILITE
-#define IMFS_DEFAULT			MFS_DEFAULT
+#define IMFS_GRAYED             MFS_GRAYED
+#define IMFS_DISABLED           MFS_DISABLED
+#define IMFS_CHECKED            MFS_CHECKED
+#define IMFS_HILITE             MFS_HILITE
+#define IMFS_ENABLED            MFS_ENABLED
+#define IMFS_UNCHECKED          MFS_UNCHECKED
+#define IMFS_UNHILITE           MFS_UNHILITE
+#define IMFS_DEFAULT            MFS_DEFAULT
 
+#endif /* WINVER >= 0x040A */
 
 /*
  * type of soft keyboard
@@ -614,21 +495,19 @@ ImmDestroySoftKeyboard(
 /* for Windows Simplified Chinese Edition */
 #define SOFTKEYBOARD_TYPE_C1            0x0002
 
-
 HIMC WINAPI ImmAssociateContext(_In_ HWND, _In_ HIMC);
+#if (WINVER >= 0x040A)
 BOOL WINAPI ImmAssociateContextEx(_In_ HWND, _In_ HIMC, _In_ DWORD);
+#endif
 
 BOOL WINAPI ImmConfigureIMEA(_In_ HKL, _In_ HWND, _In_ DWORD, _In_ LPVOID);
 BOOL WINAPI ImmConfigureIMEW(_In_ HKL, _In_ HWND, _In_ DWORD, _In_ LPVOID);
-#define ImmConfigureIME WINELIB_NAME_AW(ImmConfigureIME)
-
-HIMC WINAPI ImmCreateContext(void);
-BOOL WINAPI ImmSetActiveContext(HWND hwnd, HIMC hIMC, BOOL fFlag);
+HIMC WINAPI ImmCreateContext(VOID);
 BOOL WINAPI ImmDestroyContext(_In_ HIMC hIMC);
+#if (WINVER >= 0x040A)
 BOOL WINAPI ImmDisableIME(_In_ DWORD idThread);
 BOOL WINAPI ImmEnumInputContext(_In_ DWORD, _In_ IMCENUMPROC, _In_ LPARAM);
-BOOL WINAPI ImmLoadIME(HKL hKL);
-BOOL WINAPI CtfImmIsTextFrameServiceDisabled(VOID);
+#endif
 
 UINT
 WINAPI
@@ -650,11 +529,8 @@ ImmEnumRegisterWordW(
   _In_opt_ LPCWSTR lpszRegister,
   _In_ LPVOID);
 
-#define ImmEnumRegisterWord WINELIB_NAME_AW(ImmEnumRegisterWord)
-
 LRESULT WINAPI ImmEscapeA(_In_ HKL, _In_ HIMC, _In_ UINT, _In_ LPVOID);
 LRESULT WINAPI ImmEscapeW(_In_ HKL, _In_ HIMC, _In_ UINT, _In_ LPVOID);
-#define ImmEscape WINELIB_NAME_AW(ImmEscape)
 
 DWORD
 WINAPI
@@ -672,11 +548,8 @@ ImmGetCandidateListW(
   _Out_writes_bytes_opt_(dwBufLen) LPCANDIDATELIST lpCandList,
   _In_ DWORD dwBufLen);
 
-#define ImmGetCandidateList WINELIB_NAME_AW(ImmGetCandidateList)
-
 DWORD WINAPI ImmGetCandidateListCountA(_In_ HIMC, _Out_ LPDWORD);
 DWORD WINAPI ImmGetCandidateListCountW(_In_ HIMC, _Out_ LPDWORD);
-#define ImmGetCandidateListCount WINELIB_NAME_AW(ImmGetCandidateListCount)
 
 BOOL
 WINAPI
@@ -689,11 +562,15 @@ ImmGetCandidateWindow(
 
 BOOL WINAPI ImmGetCompositionFontA(_In_ HIMC, _Out_ LPLOGFONTA);
 BOOL WINAPI ImmGetCompositionFontW(_In_ HIMC, _Out_ LPLOGFONTW);
-#define ImmGetCompositionFont WINELIB_NAME_AW(ImmGetCompositionFont)
 
 BOOL WINAPI ImmSetCompositionFontA(_In_ HIMC, _In_ LPLOGFONTA);
 BOOL WINAPI ImmSetCompositionFontW(_In_ HIMC, _In_ LPLOGFONTW);
-#define ImmSetCompositionFont WINELIB_NAME_AW(ImmSetCompositionFont)
+
+#ifdef UNICODE
+    #define ImmGetCompositionFont ImmGetCompositionFontW
+#else
+    #define ImmGetCompositionFont ImmGetCompositionFontA
+#endif
 
 #endif
 
@@ -712,8 +589,6 @@ ImmGetCompositionStringW(
   _In_ DWORD,
   _Out_writes_bytes_opt_(dwBufLen) LPVOID lpBuf,
   _In_ DWORD dwBufLen);
-
-#define ImmGetCompositionString WINELIB_NAME_AW(ImmGetCompositionString)
 
 BOOL WINAPI ImmGetCompositionWindow(_In_ HIMC, _Out_ LPCOMPOSITIONFORM);
 HIMC WINAPI ImmGetContext(_In_ HWND);
@@ -738,8 +613,6 @@ ImmGetConversionListW(
   _In_ DWORD dwBufLen,
   _In_ UINT uFlag);
 
-#define ImmGetConversionList WINELIB_NAME_AW(ImmGetConversionList)
-
 BOOL
 WINAPI
 ImmGetConversionStatus(
@@ -763,8 +636,6 @@ ImmGetDescriptionW(
   _Out_writes_opt_(uBufLen) LPWSTR lpszDescription,
   _In_ UINT uBufLen);
 
-#define ImmGetDescription WINELIB_NAME_AW(ImmGetDescription)
-
 DWORD
 WINAPI
 ImmGetGuideLineA(
@@ -781,8 +652,6 @@ ImmGetGuideLineW(
   _Out_writes_bytes_opt_(dwBufLen) LPWSTR lpBuf,
   _In_ DWORD dwBufLen);
 
-#define ImmGetGuideLine WINELIB_NAME_AW(ImmGetGuideLine)
-
 UINT
 WINAPI
 ImmGetIMEFileNameA(
@@ -797,7 +666,7 @@ ImmGetIMEFileNameW(
   _Out_writes_opt_(uBufLen) LPWSTR lpszFileName,
   _In_ UINT uBufLen);
 
-#define ImmGetIMEFileName WINELIB_NAME_AW(ImmGetIMEFileName)
+#if (WINVER >= 0x040A)
 
 DWORD
 WINAPI
@@ -819,7 +688,13 @@ ImmGetImeMenuItemsW(
   _Out_writes_bytes_opt_(dwSize) LPIMEMENUITEMINFOW lpImeMenu,
   _In_ DWORD dwSize);
 
-#define ImmGetImeMenuItems WINELIB_NAME_AW(ImmGetImeMenuItems)
+#ifdef UNICODE
+    #define ImmGetImeMenuItems ImmGetImeMenuItemsW
+#else
+    #define ImmGetImeMenuItems ImmGetImeMenuItemsA
+#endif
+
+#endif /* (WINVER >= 0x040A) */
 
 BOOL WINAPI ImmGetOpenStatus(_In_ HIMC);
 DWORD WINAPI ImmGetProperty(_In_ HKL, _In_ DWORD);
@@ -838,24 +713,18 @@ ImmGetRegisterWordStyleW(
   _In_ UINT nItem,
   _Out_writes_(nItem) LPSTYLEBUFW lpStyleBuf);
 
-#define ImmGetRegisterWordStyle WINELIB_NAME_AW(ImmGetRegisterWordStyle)
-
 BOOL WINAPI ImmGetStatusWindowPos(_In_ HIMC, _Out_ LPPOINT);
 UINT WINAPI ImmGetVirtualKey(_In_ HWND);
 
 HKL WINAPI ImmInstallIMEA(_In_ LPCSTR, _In_ LPCSTR);
 HKL WINAPI ImmInstallIMEW(_In_ LPCWSTR, _In_ LPCWSTR);
-#define ImmInstallIME WINELIB_NAME_AW(ImmInstallIME)
 
 BOOL WINAPI ImmIsIME(_In_ HKL);
 
 BOOL WINAPI ImmIsUIMessageA(_In_ HWND, _In_ UINT, _In_ WPARAM, _In_ LPARAM);
 BOOL WINAPI ImmIsUIMessageW(_In_ HWND, _In_ UINT, _In_ WPARAM, _In_ LPARAM);
-#define ImmIsUIMessage WINELIB_NAME_AW(ImmIsUIMessage)
 
 BOOL WINAPI ImmNotifyIME(_In_ HIMC, _In_ DWORD, _In_ DWORD, _In_ DWORD_PTR);
-
-DWORD WINAPI ImmProcessKey(HWND, HKL, UINT, LPARAM, DWORD);
 
 BOOL
 WINAPI
@@ -872,8 +741,6 @@ ImmRegisterWordW(
   _In_ LPCWSTR lpszReading,
   _In_ DWORD,
   _In_ LPCWSTR lpszRegister);
-
-#define ImmRegisterWord WINELIB_NAME_AW(ImmRegisterWord)
 
 BOOL WINAPI ImmReleaseContext(_In_ HWND, _In_ HIMC);
 BOOL WINAPI ImmSetCandidateWindow(_In_ HIMC, _In_ LPCANDIDATEFORM);
@@ -898,8 +765,6 @@ ImmSetCompositionStringW(
   _Inout_updates_bytes_opt_(dwReadLen) LPVOID lpRead,
   _In_ DWORD dwReadLen);
 
-#define ImmSetCompositionString WINELIB_NAME_AW(ImmSetCompositionString)
-
 BOOL WINAPI ImmSetCompositionWindow(_In_ HIMC, _In_ LPCOMPOSITIONFORM);
 BOOL WINAPI ImmSetConversionStatus(_In_ HIMC, _In_ DWORD, _In_ DWORD);
 BOOL WINAPI ImmSetOpenStatus(_In_ HIMC, _In_ BOOL);
@@ -922,8 +787,55 @@ ImmUnregisterWordW(
   _In_ DWORD,
   _In_ LPCWSTR lpszUnregister);
 
-#define ImmUnregisterWord WINELIB_NAME_AW(ImmUnregisterWord)
-
+#ifdef UNICODE
+    #define ImmConfigureIME ImmConfigureIMEW
+    #define ImmEnumRegisterWord ImmEnumRegisterWordW
+    #define ImmEscape ImmEscapeW
+    #define ImmGetCandidateList ImmGetCandidateListW
+    #define ImmGetCandidateListCount ImmGetCandidateListCountW
+    #define ImmGetCompositionString ImmGetCompositionStringW
+    #define ImmGetConversionList ImmGetConversionListW
+    #define ImmGetDescription ImmGetDescriptionW
+    #define ImmGetGuideLine ImmGetGuideLineW
+    #define ImmGetIMEFileName ImmGetIMEFileNameW
+    #define ImmGetRegisterWordStyle ImmGetRegisterWordStyleW
+    #define ImmInstallIME ImmInstallIMEW
+    #define ImmIsUIMessage ImmIsUIMessageW
+    #define ImmRegisterWord ImmRegisterWordW
+    #define ImmRequestMessage ImmRequestMessageW
+    #define ImmSetCompositionFont ImmSetCompositionFontW
+    #define ImmSetCompositionString ImmSetCompositionStringW
+    #define ImmUnregisterWord ImmUnregisterWordW
+    typedef LPSTYLEBUFW LPSTYLEBUF;
+    typedef NPSTYLEBUFW NPSTYLEBUF;
+    typedef PSTYLEBUFW PSTYLEBUF;
+    typedef REGISTERWORDENUMPROCW REGISTERWORDENUMPROC;
+    typedef STYLEBUFW STYLEBUF;
+#else
+    #define ImmConfigureIME ImmConfigureIMEA
+    #define ImmEnumRegisterWord ImmEnumRegisterWordA
+    #define ImmEscape ImmEscapeA
+    #define ImmGetCandidateList ImmGetCandidateListA
+    #define ImmGetCandidateListCount ImmGetCandidateListCountA
+    #define ImmGetCompositionString ImmGetCompositionStringA
+    #define ImmGetConversionList ImmGetConversionListA
+    #define ImmGetDescription ImmGetDescriptionA
+    #define ImmGetGuideLine ImmGetGuideLineA
+    #define ImmGetIMEFileName ImmGetIMEFileNameA
+    #define ImmGetRegisterWordStyle ImmGetRegisterWordStyleA
+    #define ImmInstallIME ImmInstallIMEA
+    #define ImmIsUIMessage ImmIsUIMessageA
+    #define ImmRegisterWord ImmRegisterWordA
+    #define ImmRequestMessage ImmRequestMessageA
+    #define ImmSetCompositionFont ImmSetCompositionFontA
+    #define ImmSetCompositionString ImmSetCompositionStringA
+    #define ImmUnregisterWord ImmUnregisterWordA
+    typedef LPSTYLEBUFA LPSTYLEBUF;
+    typedef NPSTYLEBUFA NPSTYLEBUF;
+    typedef PSTYLEBUFA PSTYLEBUF;
+    typedef REGISTERWORDENUMPROCA REGISTERWORDENUMPROC;
+    typedef STYLEBUFA STYLEBUF;
+#endif
 
 #ifdef __cplusplus
 }
