+++ win32ss/user/user32/windows/defwnd.c
@@ -10,7 +10,7 @@
  */
 
 #include <user32.h>
-#include <ddk/immdev.h>
+#include <immdev.h>
 
 WINE_DEFAULT_DEBUG_CHANNEL(user32);
