+++ win32ss/user/winsrv/winsrv.h
@@ -23,6 +23,8 @@
 #include <winreg.h>
 #include <winuser.h>
 #include <imm.h>
+#include <immdev.h>
+#include <imm32_undoc.h>
 
 /* Undocumented user definitions */
 #include <undocuser.h>
